+++ packages/tonal/test/tonal-test.js
@@ -3,7 +3,7 @@ var tonal = require('..')
 
 tape('tonal', function (test) {
   test.ok(tonal, 'Tonal object')
-  test.equal(Object.keys(tonal).length, 27, 'tonal exports count')
+  test.equal(Object.keys(tonal).length, 31, 'tonal exports count')
   Object.keys(tonal).forEach(function (name) {
     test.ok(tonal[name], 'Function ' + name)
   })